fix(mac): hide menu bar limit metrics for providers hidden in limits display
Hiding a provider in Limits Display preferences now also removes its
metrics from the menu bar, completing the #162 repro path:

- availableItemIDs gains a hiddenProviders filter that excludes hidden
  providers' metrics even when currently selected, so pushSettings
  self-heals the stored selection and the dashboard dropdown drops them.
- buildMenuBarDisplayValues filters at render time as well, because a
  dashboard-applied snapshot re-renders via .nativeSettingsChanged before
  the stored selection is re-normalized.
- Applied bridge snapshots now fall through to pushSettings instead of
  returning early; the echo is convergent (the dashboard receives its own
  snapshot and does not write back) and the push refreshes menu bar
  candidates after visibility changes.
- Extract MenuBarDisplayMetric/MenuBarDisplayPreferences from
  StatusBarController into Models/MenuBarDisplayPreferences.swift so the
  unit-test target can compile them, and add coverage for the hidden
  filter plus a Swift<->JS provider id parity guardrail.
